Revert "Theme CSS to JSON Convertion: All Chat Components & Page"
