+++ src/javascript/components/Chat/AddStickerPopup.tsx
@@ -3,19 +3,24 @@ import { useContext, Component, useState, useEffect } from 'react';
 import { theme, ThemeContext } from '../../helpers';
 import { MdSend, MdPerson, MdMood, MdFace } from 'react-icons/md';
 
+import { Message } from './Message';
 import { firebaseEmotes$, setRxEmotes } from '../../helpers/rxEmotes';
 import { Action } from 'rxjs/internal/scheduler/Action';
 
 import { SegmentControl, SegmentControlSource } from '../SegmentControl/index';
-import { Button, DestructiveButton, ActionButton, Message } from '../Generics/CreativeUI';
+import { Button, DestructiveButton, ActionButton } from '../Generics/Button';
 
 const Window: any = window;
 const { ipcRenderer, shell } = Window.require('electron');
+
+const styles: any = require('./Chat.scss');
+const segStyles: any = require('../SegmentControl/SegmentControl.scss');
 
 interface popup {
   stickerUrl: String;
   stickerDLiveId: String;
   stickerId: String;
+  styles: any;
   stateTheme: any;
   text?: string | Function | Element | any;
   Config?: any;
@@ -27,6 +32,7 @@ const AddStickerPopup = ({
   stickerId,
   stickerDLiveId,
   stickerUrl,
+  styles,
   stateTheme,
   text = '',
   Config = {},
@@ -92,41 +98,41 @@ const AddStickerPopup = ({
   };
 
   return (
-    <div style={stateTheme.popup.dialog.content}>
+    <div className={`${styles.popup}`}>
       <h2>Add Sticker</h2>
-      <div style={stateTheme.popup.dialog.content.fullWidth}>
+      <div className={`${styles.stickersPopup}`}>
         <div>
-          <div style={stateTheme.popup.dialog.content.addSticker.imageContainer}>
-            <img  style={stateTheme.popup.dialog.content.addSticker.imageContainer.image} src={`${stickerUrl}`} />
+          <div className={styles.sticker_container}>
+            <img className={`${styles.previewSticker}`} src={`${stickerUrl}`} />
           </div>
-          <div style={stateTheme.popup.dialog.content.addSticker.idContainer}>
-            <span style={stateTheme.popup.dialog.content.addSticker.idContainer.title}>Sticker DLive Id:</span>
-            <span style={stateTheme.popup.dialog.content.addSticker.idContainer.info}>{cleanId()}</span>
+          <div className={styles.stickerIdContainer}>
+            <span className={styles.idTitle}>Sticker DLive Id:</span>
+            <span className={styles.idInfo}>{cleanId()}</span>
           </div>
         </div>
-        <div style={stateTheme.popup.dialog.buttonStack}>
-          <Button
-            title={'Send Sticker'}
-            isSubmit={true}
-            stateTheme={stateTheme}
-            buttonStyle={{ width: '-webkit-fill-available' }}
-            onClick={() => {
-              sendMessage();
-            }}
-          />
-          <Button
-            title={'Add Sticker'}
-            isSubmit={true}
-            stateTheme={stateTheme}
-            buttonStyle={{
-              width: '-webkit-fill-available',
-              'margin-left': '10px'
-            }}
-            onClick={() => {
-              save();
-            }}
-          />
-        </div>
+      </div>
+      <div className={styles.buttonstack}>
+        <Button
+          title={'Send Sticker'}
+          isSubmit={true}
+          stateTheme={stateTheme}
+          buttonStyle={{ width: '-webkit-fill-available' }}
+          onClick={() => {
+            sendMessage();
+          }}
+        />
+        <Button
+          title={'Add Sticker'}
+          isSubmit={true}
+          stateTheme={stateTheme}
+          buttonStyle={{
+            width: '-webkit-fill-available',
+            'margin-left': '10px'
+          }}
+          onClick={() => {
+            save();
+          }}
+        />
       </div>
     </div>
   );
