+++ src/javascript/components/Chat/ChatEventsPopup.tsx
@@ -3,30 +3,33 @@ import { useContext, Component, useState, useEffect } from 'react';
 import { theme, ThemeContext } from '../../helpers';
 import { MdSend, MdPerson, MdMood, MdFace, MdPlayArrow } from 'react-icons/md';
 
+import { Message } from './Message';
 import { firebaseConfig$, setRxConfig } from '../../helpers/rxConfig';
 import { Action } from 'rxjs/internal/scheduler/Action';
 
 // Generic Components
-import { 
-  Toggle, 
-  ToggleType,
-  Panel,
+import { SegmentControl, SegmentControlSource } from '../SegmentControl/index';
+import { Toggle, ToggleType } from '../Generics/Toggle';
+import { Panel } from '../Generics/Panel';
+import {
   Button,
   DestructiveButton,
   ActionButton,
   WidgetButton,
-  IconButton,
-  TextField, 
-  StepperField,
-  ScrollView 
-  } from '../Generics/CreativeUI';
+  SendButton
+} from '../Generics/Button';
+import { TextField, StepperField } from '../Generics/Input';
 import { GoNoNewline } from 'react-icons/go';
 import { first } from 'rxjs/operators';
 
 const Window: any = window;
 const { ipcRenderer, shell } = Window.require('electron');
 
+const styles: any = require('./Chat.scss');
+const segStyles: any = require('../SegmentControl/SegmentControl.scss');
+
 interface popup {
+  styles: any;
   stateTheme: any;
   text?: string | Function | Element | any;
   Config?: any;
@@ -43,6 +46,7 @@ let ninjaTimeout;
 let ninjetTimeout;
 
 const ChatEventsPopup = ({
+  styles,
   stateTheme,
   text = '',
   Config = {},
@@ -215,352 +219,339 @@ const ChatEventsPopup = ({
   };
 
   return (
-    <div style={stateTheme.popup.dialog.content}>
+    <div className={`${styles.popup}`}>
       <h2>Chat on Events</h2>
-      <div style={stateTheme.popup.dialog.content.fullWidth}>
-        <ScrollView stateTheme={stateTheme}>
-          <Toggle
-            header='Send Event Messages'
-            type={ToggleType.stretched}
-            isEnabled={true}
-            isOn={enableEventMessages}
-            onClick={() => {
-              updateEnableEvents(enableEventMessages);
-            }}
-            stateTheme={stateTheme}
-          />
-
-          <Panel
-            header={enableEventMessages ? null : 'Event Messages Disabled'}
-            hasHeader={!enableEventMessages}
-            style={stateTheme.base.tertiaryBackground}
-            stateTheme={stateTheme}
-            content={
-              enableEventMessages ? (
-                <div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'On Follow'}
-                        placeholderText='No message...'
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        text={onFollow}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateFollow(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'On Sub'}
-                        placeholderText='No message...'
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        text={onSub}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateSub(e.target.value);
-                          // if (!isNaN(Number(e.target.value)) || e.target.value === '')
-                          //   setPoints(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'On Gift Sub'}
-                        placeholderText='No message...'
-                        stateTheme={stateTheme}
-                        text={onGiftedSub}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateGiftedSub(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                </div>
-              ) : null
-            }
-          />
-          <Panel
-            header={
-              enableEventMessages
-                ? 'Donation Messages'
-                : 'Event Messages Disabled'
-            }
-            hasHeader={true}
-            style={Object.assign(
-              {},
-              { marginTop: '10px' },
-              stateTheme.base.tertiaryBackground
-            )}
-            stateTheme={stateTheme}
-            content={
-              enableEventMessages ? (
-                <div>
-                  <Toggle
-                    header='Spam Filter'
-                    type={ToggleType.stretched}
-                    isEnabled={true}
-                    isOn={enableDebounceEvents}
-                    onClick={() => {
-                      updateDebounceEvents(enableDebounceEvents);
-                    }}
-                    stateTheme={stateTheme}
-                  />
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        placeholderText='No message...'
-                        header={'Lemon'}
-                        stateTheme={stateTheme}
-                        text={onLemon}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateLemon(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'Icecream'}
-                        placeholderText='No message...'
-                        text={onIcecream}
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateIcecream(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'Diamond'}
-                        placeholderText='No message...'
-                        text={onDiamond}
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateDiamond(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'Ninjaghini'}
-                        placeholderText='No message...'
-                        text={onNinja}
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateNinja(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <div style={{ flex: 1 }}>
-                      <TextField
-                        header={'Ninjet'}
-                        placeholderText='No message...'
-                        text={onNinjet}
-                        stateTheme={stateTheme}
-                        width={'100%'}
-                        inputStyle={stateTheme.base.secondaryBackground}
-                        onChange={e => {
-                          updateNinjet(e.target.value);
-                        }}
-                      />
-                    </div>
-                    <IconButton
-                      icon={
-                        <MdPlayArrow style={{ height: '30px  !important', width: '30px' }} />
-                      }
-                      stateTheme={stateTheme}
-                      buttonStyle={{
-                        marginTop: '30px',
-                        marginBottom: '10px',
-                        maxHeight: '30px',
-                        height: '30px',
-                      }}
-                      onClick={() => {}}
-                    />
-                  </div>
-                </div>
-              ) : null
-            }
-          />
-        </ScrollView>
-        <div style={stateTheme.popup.dialog.buttonStack}>
-          <Button
-          title={'Close'}
-          isSubmit={true}
-          buttonStyle={{ width: '100%'}}
-          stateTheme={stateTheme}
+      <div className={`${styles.chatFilterPopup}`}>
+        <Toggle
+          header='Send Event Messages'
+          type={ToggleType.stretched}
+          isEnabled={true}
+          isOn={enableEventMessages}
           onClick={() => {
-            setTimeout(() => {
-              closeCurrentPopup();
-            }, 400);
+            updateEnableEvents(enableEventMessages);
           }}
+          stateTheme={stateTheme}
         />
-        </div>
+
+        <Panel
+          header={enableEventMessages ? null : 'Event Messages Disabled'}
+          hasHeader={!enableEventMessages}
+          style={stateTheme.base.tertiaryBackground}
+          stateTheme={stateTheme}
+          content={
+            enableEventMessages ? (
+              <div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'On Follow'}
+                      placeholderText='No message...'
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      text={onFollow}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateFollow(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'On Sub'}
+                      placeholderText='No message...'
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      text={onSub}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateSub(e.target.value);
+                        // if (!isNaN(Number(e.target.value)) || e.target.value === '')
+                        //   setPoints(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'On Gift Sub'}
+                      placeholderText='No message...'
+                      stateTheme={stateTheme}
+                      text={onGiftedSub}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateGiftedSub(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+              </div>
+            ) : null
+          }
+        />
+        <Panel
+          header={
+            enableEventMessages
+              ? 'Donation Messages'
+              : 'Event Messages Disabled'
+          }
+          hasHeader={true}
+          style={Object.assign(
+            {},
+            { marginTop: '10px' },
+            stateTheme.base.tertiaryBackground
+          )}
+          stateTheme={stateTheme}
+          content={
+            enableEventMessages ? (
+              <div>
+                <Toggle
+                  header='Spam Filter'
+                  type={ToggleType.stretched}
+                  isEnabled={true}
+                  isOn={enableDebounceEvents}
+                  onClick={() => {
+                    updateDebounceEvents(enableDebounceEvents);
+                  }}
+                  stateTheme={stateTheme}
+                />
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      placeholderText='No message...'
+                      header={'Lemon'}
+                      stateTheme={stateTheme}
+                      text={onLemon}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateLemon(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'Icecream'}
+                      placeholderText='No message...'
+                      text={onIcecream}
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateIcecream(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'Diamond'}
+                      placeholderText='No message...'
+                      text={onDiamond}
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateDiamond(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'Ninjaghini'}
+                      placeholderText='No message...'
+                      text={onNinja}
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateNinja(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+                <div
+                  style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center'
+                  }}
+                >
+                  <div style={{ flex: 1 }}>
+                    <TextField
+                      header={'Ninjet'}
+                      placeholderText='No message...'
+                      text={onNinjet}
+                      stateTheme={stateTheme}
+                      width={'100%'}
+                      inputStyle={stateTheme.base.secondaryBackground}
+                      onChange={e => {
+                        updateNinjet(e.target.value);
+                      }}
+                    />
+                  </div>
+                  <SendButton
+                    icon={
+                      <MdPlayArrow style={{ height: '25px', width: '25px' }} />
+                    }
+                    stateTheme={stateTheme}
+                    buttonStyle={{
+                      marginTop: '21px',
+                      marginBottom: '10px',
+                      maxHeight: '30px'
+                    }}
+                    onClick={() => {}}
+                  />
+                </div>
+              </div>
+            ) : null
+          }
+        />
       </div>
+      <Button
+        title={'Close'}
+        isSubmit={true}
+        stateTheme={stateTheme}
+        onClick={() => {
+          setTimeout(() => {
+            closeCurrentPopup();
+          }, 400);
+        }}
+      />
     </div>
   );
 };
