+++ src/javascript/components/Chat/ChatTextToSpeechPopup.tsx
@@ -3,6 +3,7 @@ import { useContext, Component, useState, useEffect } from 'react';
 import { theme, ThemeContext } from '../../helpers';
 import { MdSend, MdPerson, MdMood, MdFace } from 'react-icons/md';
 
+import { Message } from './Message';
 import { firebaseConfig$, setRxConfig } from '../../helpers/rxConfig';
 import { Action } from 'rxjs/internal/scheduler/Action';
 
@@ -16,7 +17,11 @@ import { Button, DestructiveButton, ActionButton } from '../Generics/Button';
 const Window: any = window;
 const { ipcRenderer, shell } = Window.require('electron');
 
+const styles: any = require('./Chat.scss');
+const segStyles: any = require('../SegmentControl/SegmentControl.scss');
+
 interface popup {
+  styles: any;
   stateTheme: any;
   text?: string | Function | Element | any;
   Config?: any;
@@ -26,6 +31,7 @@ interface popup {
 let timeout;
 
 const ChatTextToSpeechPopup = ({
+  styles,
   stateTheme,
   text = '',
   Config = {},
@@ -73,9 +79,9 @@ const ChatTextToSpeechPopup = ({
   };
 
   return (
-    <div style={stateTheme.popup.dialog.content}>
+    <div className={`${styles.popup}`}>
       <h2>Text To Speech Settings</h2>
-      <div style={stateTheme.popup.dialog.content.seventyWidth}>
+      <div className={`${styles.chatFilterPopup}`}>
         <Toggle
           header='Enable TTS on Donations'
           type={ToggleType.stretched}
@@ -95,6 +101,7 @@ const ChatTextToSpeechPopup = ({
             setTTSAmplitude(value);
             saveToDB('tts_Amplitude', value);
           }}
+          style={styles}
           stateTheme={stateTheme}
         />
         <RangeSlider
@@ -106,6 +113,7 @@ const ChatTextToSpeechPopup = ({
             setTTSPitch(value);
             saveToDB('tts_Pitch', value);
           }}
+          style={styles}
           stateTheme={stateTheme}
         />
         <RangeSlider
@@ -117,11 +125,12 @@ const ChatTextToSpeechPopup = ({
             setTTSSpeed(value);
             saveToDB('tts_Speed', value);
           }}
+          style={styles}
           stateTheme={stateTheme}
         />
         {/*<Slider header="Word Gap" val={ttsWordGap} valType={"ms"} maxValue={100} hasHeader={true} onChange={(e, value) => {setTTSWordGap(value); saveToDB("tts_WordGap", value);}} style={styles}/>*/}
       </div>
-      <div style={stateTheme.popup.dialog.buttonStack}>
+      <div className={styles.buttonstack}>
         <Button
           title={'Test TTS'}
           isSubmit={true}
