+++ src/javascript/components/Chat/Emote.scss
@@ -1,0 +1,64 @@
+@import './../../../styles/variables.scss';
+
+.emote_container {
+    display: inline-block;
+    width: 100px;
+    height: 100px;
+    margin-right: 10px;
+    margin-bottom: 10px;
+    box-sizing: border-box;
+    position: relative;
+
+    .emoteDeleteButton { display: none; }
+
+    &:hover {
+        .emoteDeleteButton {
+            display: block;
+            position: absolute;
+            background: white;
+            width: 20px;
+            height: 20px;
+            border-radius: 10px;
+            float: right;
+            top: 0;
+            right: 0;
+            box-shadow: 1px 1px 30px rgba(21,21,21,0.6);
+
+            &:hover {
+                cursor: pointer;
+                svg {
+                    fill: red !important;
+                    stroke: red !important;
+                }
+            }
+
+            svg {
+                vertical-align: middle;
+                text-align: center;
+                height: 14px;
+                width: 14px;
+                margin: 0px;
+                margin-bottom: 3px;
+                margin-left: 0px;
+                fill: Black !important;
+                stroke: black !important;
+
+                &:hover {
+                    fill: red !important;
+                    stroke: red !important;
+                }
+            }
+        }
+        .emote {
+            border: 4px solid $dliveaccent;
+            border-radius: 8px;
+        }
+    }
+
+    .emote {
+        box-sizing: border-box;
+        width: 100px;
+        max-height: 100px;
+        user-select: none;
+    }
+}
