+++ src/javascript/components/Chat/StickerPopup.tsx
@@ -10,23 +10,25 @@ import {
   MdSentimentDissatisfied
 } from 'react-icons/md';
 
-import { ScrollView } from '../Generics/CreativeUI';
-
+import { Message } from './Message';
 import { firebaseEmotes$, setRxEmotes } from '../../helpers/rxEmotes';
 import { Action } from 'rxjs/internal/scheduler/Action';
 
 import { SegmentControl, SegmentControlSource } from '../SegmentControl/index';
 import { SegmentControlItem } from '../SegmentControl/SegmentControlItem';
-import { Button, DestructiveButton, ActionButton, Emote } from '../Generics/CreativeUI';
+import { Button, DestructiveButton, ActionButton } from '../Generics/Button';
 
+import { Emote } from './Emote';
 import { string } from 'prop-types';
 
 const Window: any = window;
 const { ipcRenderer, shell } = Window.require('electron');
 
+const styles: any = require('./Chat.scss');
 const segStyles: any = require('../SegmentControl/SegmentControl.scss');
 
 interface popup {
+  styles: any;
   stateTheme: any;
   text?: string | Function | Element | any;
   Config?: any;
@@ -34,6 +36,7 @@ interface popup {
 }
 
 const StickerPopup = ({
+  styles,
   stateTheme,
   text = '',
   Config = {},
@@ -154,9 +157,9 @@ const StickerPopup = ({
   };
 
   return (
-    <div style={stateTheme.popup.dialog.content}>
+    <div className={styles.popup}>
       <h2>Stickers</h2>
-      <div style={stateTheme.popup.dialog.content.fullWidth}>
+      <div className={`${styles.stickersPopup}`}>
         <div className={segStyles.segmentControl}>
           <div
             className={segStyles.segmentHeader}
@@ -173,44 +176,23 @@ const StickerPopup = ({
           </div>
           <div className={segStyles.segmentBody}>
             <div className={segStyles.segmentView}>
-              <ScrollView stateTheme={stateTheme}>
-              {/*<Emote
-                  stickerDLiveId={':emote/mine/dlive-53093718/36d1544a90081e3_300300:'}
-                  stickerUrl={'https://images.prd.dlivecdn.com/emote/37a2e5297000955_288300'}
-                  canDelete={false}
-                  />
-                  <Emote
-                  stickerDLiveId={':emote/mine/dlive-53093718/36d1544a90081e3_300300:'}
-                  stickerUrl={'https://images.prd.dlivecdn.com/emote/37a2e5297000955_288300'}
-                  canDelete={false}
-                  />
-                  <Emote
-                  stickerDLiveId={':emote/mine/dlive-53093718/36d1544a90081e3_300300:'}
-                  stickerUrl={'https://images.prd.dlivecdn.com/emote/37a2e5297000955_288300'}
-                  canDelete={false}
-                  />
-                  <Emote
-                  stickerDLiveId={':emote/mine/dlive-53093718/36d1544a90081e3_300300:'}
-                  stickerUrl={'https://images.prd.dlivecdn.com/emote/37a2e5297000955_288300'}
-                  canDelete={false}
-                />*/}
+              <div className={styles.gridView}>
                 {emotesList.length > 0 ? (
                   emotesList.map(i => (
                     <Emote
                       stickerDLiveId={i.dliveid}
                       stickerUrl={i.url}
-                      stateTheme={stateTheme}
                       canDelete={index == 1 ? true : false}
                       onDelete={() => deleteEmote(i.id)}
                     />
                   ))
                 ) : (
-                  <div style={stateTheme.chatPage.stickers.noStickers}>
-                    <MdSentimentDissatisfied  style={stateTheme.chatPage.stickers.noStickers.svg}/>
+                  <div className={styles.noStickers}>
+                    <MdSentimentDissatisfied />
                     <h3>{noStickerErrorMsg[index]}</h3>
                   </div>
                 )}
-              </ScrollView>
+              </div>
             </div>
           </div>
         </div>
