+++ src/javascript/components/Generics/Button.tsx
@@ -140,13 +140,12 @@ const ActionButton = ({
   );
 };
 
-const IconButton = ({
+const SendButton = ({
   icon,
   width = 'auto',
   isEnabled = true,
   onClick = null,
   stateTheme,
-  style = {},
   buttonStyle = {}
 }) => {
   return (
@@ -156,25 +155,23 @@ const IconButton = ({
         !isEnabled ? stateTheme.button.normal.disabled : null,
         Object.assign(
           {},
-          Object.assign({}, stateTheme.button.sender, style),
           Object.assign({}, stateTheme.base.quinaryBackground, {
             borderColor: 'transparent'
-          })
+          }),
+          stateTheme.button.sender
         )
       )}
       hoverStyle={Object.assign(
         {},
         Object.assign({}, theme.globals.accentFillColor, {
-          borderColor: theme.globals.accentBorderColor.borderColor,
-          fill: theme.globals.accentBackground.backgroundColor 
+          borderColor: theme.globals.accentBorderColor.borderColor
         }),
         stateTheme.button.sender.hover
       )}
       isButton={true}
       aStyle={buttonStyle}
-      onClick={onClick}
     >
-      {icon}
+      <div onClick={onClick}>{icon}</div>
     </AdvancedDiv>
   );
 };
@@ -184,8 +181,6 @@ const WidgetButton = ({
   style = null,
   isEnabled = true,
   onClick = null,
-  header = null,
-  footer = null,
   stateTheme
 }) => {
   const [isenabled, setIsEnabled] = useState(isEnabled);
@@ -199,51 +194,13 @@ const WidgetButton = ({
       )}
       hoverStyle={Object.assign(
         {},
-        { 
-          cursor: 'pointer'
-        },
+        { cursor: 'pointer' },
         theme.globals.accentFillColor
       )}
       isButton={true}
       aStyle={Object.assign({}, stateTheme.button.widget, style)}
     >
-      <div 
-        onClick={onClick} 
-        style={{
-          'verticalAlign': 'middle',
-          'display': 'flex'
-      }}>
-        {header != null ? header : null}
-        {icon}
-        {footer != null ? footer : null}
-      </div>
-    </AdvancedDiv>
-  );
-};
-
-const BubbleButton = ({
-  icon,
-  hoverTextColor = 'green',
-  style = null,
-  isEnabled = true,
-  onClick = null,
-  header = null,
-  footer = null,
-  stateTheme
-}) => {
-  const [isenabled, setIsEnabled] = useState(isEnabled);
-
-  return (
-    <AdvancedDiv 
-      aStyle={stateTheme.button.bubble} 
-      style={stateTheme.button.bubble.svg}
-      hoverStyle={Object.assign({}, { 
-        fill: hoverTextColor, 
-        stroke: hoverTextColor, 
-        color: hoverTextColor} ,
-        stateTheme.button.bubble.hover.svg)}
-      onClick={onClick}>
-      {icon}
+      <div onClick={onClick}>{icon}</div>
     </AdvancedDiv>
   );
 };
@@ -295,8 +252,7 @@ export {
   Button,
   DestructiveButton,
   ActionButton,
-  IconButton,
+  SendButton,
   WidgetButton,
-  BubbleButton,
   LinkButton
 };
