+++ src/javascript/components/Generics/Common.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import { useState } from 'react';
 
+import { AdvancedDiv } from '../Generics/AdvancedDiv';
+
 const Page = ({ children, stateTheme, style = {} }) => {
 
   return (
