+++ src/styles/main.css
@@ -6,10 +6,9 @@ body {
   overflow: hidden;
 }
 svg:hover {
-  /*fill: #ffd300 !important;
-  stroke: #ffd300 !important;*/
+  fill: #ffd300 !important;
+  stroke: #ffd300 !important;
 }
-
 #app {
   height: 100%;
   width: 100%;
